gemini_live_bridge/requirements.txt
google-genai>=0.3.0
openai>=1.0.0
aiohttp>=3.9.0
numpy>=1.24.0
